requirements.txt
gradio==3.49.0
numpy==1.24.2
openai==1.10.0
pandas==2.1.2
PyYAML==6.0.1
Telethon==1.34.0